stub in Chart, define platforms
platforms stanza added to Package.swift based on wanting to leverage
SwiftUI Canvas, so availability matches.

Chart stubbed out as a struct conforming to view, with an initial
initializer that takes a standard ViewBuilder to generate the content.
This will get updated based on what happens with Mark - if that returns
a struct that is assembled into a View using Canvas, or if each Mark
returns a View that's composed by Chart.