rough out the internals of VisualChannel
Dug into KeyPath's sufficient to have a pattern that can work with
Swift's type system to map values from one type to another through
a VisualChannel. Got far enough down to wanting to use it within a
result builder to realize that I likely need to at least partially type
erase the full signature type in order to use it as the result of a
build block.